Update README.md
Updated the title to Dawn of Djinn minus since I've gotten a handful of complaints of QOL not delivering the objectively best experience. DoD Minus should better reflect this being a mod that offers DoD features in the most vanilla way possible, and not nearly the defacto Golden Sun experience. 

I've also updated the FAQ to include answers to the lack of Bilibin Barricade & addressing a technical oversight with that comes from fast traveling to Vault before completing Lamakan Desert.

Please stop asking about Bilibin Barricade I will not be updating the fast travel to include it. Everything here is either taken from the GS Randomizer or made in a simple UI editor. I can't change or remove that code, even if I wanted to.